lib/{vfscore,posix-fdtab}: Split fd table mechanism into posix-fdtab
This moves the fd table handling into its own library.

Checkpatch-Ignore: USE_NEGATIVE_ERRNO
Checkpatch-Ignore: NON_RAW_SYSCALL
GitHub-Fixes: #52
