+++ lib/Makefile.uk
@@ -9,6 +9,7 @@ $(eval $(call _import_lib,$(CONFIG_UK_BASE)/lib/devfs))
 $(eval $(call _import_lib,$(CONFIG_UK_BASE)/lib/fdt))
 $(eval $(call _import_lib,$(CONFIG_UK_BASE)/lib/isrlib))
 $(eval $(call _import_lib,$(CONFIG_UK_BASE)/lib/nolibc))
+$(eval $(call _import_lib,$(CONFIG_UK_BASE)/lib/posix-fdtab))
 $(eval $(call _import_lib,$(CONFIG_UK_BASE)/lib/posix-event))
 $(eval $(call _import_lib,$(CONFIG_UK_BASE)/lib/posix-libdl))
 $(eval $(call _import_lib,$(CONFIG_UK_BASE)/lib/posix-mmap))
