+++ lib/posix-fdtab/Makefile.uk
@@ -1,0 +1,27 @@
+$(eval $(call addlib_s,libposix-fdtab,$(CONFIG_LIBPOSIX_FDTAB)))
+
+CINCLUDES-y += -I$(LIBPOSIX-FDTAB_BASE)/include
+
+LIBPOSIX-FDTAB_SRCS-y += $(LIBPOSIX-FDTAB_BASE)/fd.c
+LIBPOSIX-FDTAB_SRCS-y += $(LIBPOSIX-FDTAB_BASE)/main.c
+LIBPOSIX-FDTAB_SRCS-y += $(LIBPOSIX-FDTAB_BASE)/syscalls.c
+LIBPOSIX-FDTAB_SRCS-y += $(LIBPOSIX-FDTAB_BASE)/subr_uio.c
+LIBPOSIX-FDTAB_SRCS-y += $(LIBPOSIX-FDTAB_BASE)/eventpoll.c
+
+UK_PROVIDED_SYSCALLS-$(CONFIG_LIBPOSIX_FDTAB) += write-3 writev-3 pwrite64-4
+UK_PROVIDED_SYSCALLS-$(CONFIG_LIBPOSIX_FDTAB) += read-3 readv-3 pread64-4
+UK_PROVIDED_SYSCALLS-$(CONFIG_LIBPOSIX_FDTAB) += fstat-2
+UK_PROVIDED_SYSCALLS-$(CONFIG_LIBPOSIX_FDTAB) += fcntl-3
+UK_PROVIDED_SYSCALLS-$(CONFIG_LIBPOSIX_FDTAB) += close-1
+UK_PROVIDED_SYSCALLS-$(CONFIG_LIBPOSIX_FDTAB) += lseek-3
+UK_PROVIDED_SYSCALLS-$(CONFIG_LIBPOSIX_FDTAB) += ftruncate-2
+UK_PROVIDED_SYSCALLS-$(CONFIG_LIBPOSIX_FDTAB) += fallocate-4
+UK_PROVIDED_SYSCALLS-$(CONFIG_LIBPOSIX_FDTAB) += dup-1
+UK_PROVIDED_SYSCALLS-$(CONFIG_LIBPOSIX_FDTAB) += dup3-3
+UK_PROVIDED_SYSCALLS-$(CONFIG_LIBPOSIX_FDTAB) += dup2-2
+UK_PROVIDED_SYSCALLS-$(CONFIG_LIBPOSIX_FDTAB) += pwritev-4
+UK_PROVIDED_SYSCALLS-$(CONFIG_LIBPOSIX_FDTAB) += fsync-1
+UK_PROVIDED_SYSCALLS-$(CONFIG_LIBPOSIX_FDTAB) += fdatasync-1
+UK_PROVIDED_SYSCALLS-$(CONFIG_LIBPOSIX_FDTAB) += preadv-4
+UK_PROVIDED_SYSCALLS-$(CONFIG_LIBPOSIX_FDTAB) += flock-2
+UK_PROVIDED_SYSCALLS-$(CONFIG_LIBPOSIX_FDTAB) += ioctl-3
