+++ lib/posix-fdtab/exportsyms.uk
@@ -1,0 +1,94 @@
+fdtab_get_active
+fdtab_set_active
+fdtab_reserve_fd
+fdtab_fdrop
+fdtab_fhold
+fdtab_alloc_fd
+fdtab_put_fd
+fdtab_install_fd
+fdtab_file_init
+fdtab_get_file
+fdtab_put_file
+fdtab_fdalloc
+fdtab_fget
+fdtab_uiomove
+write
+uk_syscall_e_write
+uk_syscall_r_write
+close
+uk_syscall_e_close
+uk_syscall_r_close
+read
+uk_syscall_e_read
+uk_syscall_r_read
+fsync
+uk_syscall_e_fsync
+uk_syscall_r_fsync
+dup
+uk_syscall_e_dup
+uk_syscall_r_dup
+dup2
+uk_syscall_e_dup2
+uk_syscall_r_dup2
+dup3
+uk_syscall_e_dup3
+uk_syscall_r_dup3
+fstat
+fstat64
+uk_syscall_e_fstat
+uk_syscall_r_fstat
+fcntl
+uk_syscall_e_fcntl
+uk_syscall_r_fcntl
+fdatasync
+uk_syscall_e_fdatasync
+uk_syscall_r_fdatasync
+uk_syscall_e_writev
+uk_syscall_r_writev
+flock
+uk_syscall_e_flock
+uk_syscall_r_flock
+lseek
+lseek64
+uk_syscall_e_lseek
+uk_syscall_r_lseek
+preadv
+preadv64
+uk_syscall_e_preadv
+uk_syscall_r_preadv
+ftruncate
+ftruncate64
+uk_syscall_e_ftruncate
+uk_syscall_r_ftruncate
+fallocate
+fallocate64
+uk_syscall_e_fallocate
+uk_syscall_r_fallocate
+ioctl
+uk_syscall_e_ioctl
+uk_syscall_r_ioctl
+pread
+pread64
+uk_syscall_e_pread64
+uk_syscall_r_pread64
+pwritev
+pwritev64
+uk_syscall_e_pwritev
+uk_syscall_r_pwritev
+pwrite
+pwrite64
+uk_syscall_e_pwrite64
+uk_syscall_r_pwrite64
+readv
+uk_syscall_e_readv
+uk_syscall_r_readv
+writev
+eventpoll_fini
+eventpoll_notify_close
+eventpoll_add
+eventpoll_add_unsafe
+eventpoll_mod
+eventpoll_del
+eventpoll_del_unsafe
+eventpoll_wait
+eventpoll_signal
