+++ lib/vfscore/Makefile.uk
@@ -16,52 +16,33 @@ LIBVFSCORE_SRCS-y += $(LIBVFSCORE_BASE)/main.c
 LIBVFSCORE_SRCS-y += $(LIBVFSCORE_BASE)/task.c
 LIBVFSCORE_SRCS-y += $(LIBVFSCORE_BASE)/lookup.c
 LIBVFSCORE_SRCS-y += $(LIBVFSCORE_BASE)/fops.c
-LIBVFSCORE_SRCS-y += $(LIBVFSCORE_BASE)/subr_uio.c
 LIBVFSCORE_SRCS-y += $(LIBVFSCORE_BASE)/pipe.c
-LIBVFSCORE_SRCS-y += $(LIBVFSCORE_BASE)/eventpoll.c
 LIBVFSCORE_SRCS-y += $(LIBVFSCORE_BASE)/extra.ld
 LIBVFSCORE_SRCS-$(CONFIG_LIBVFSCORE_AUTOMOUNT_ROOTFS) += \
 	$(LIBVFSCORE_BASE)/rootfs.c
 
 
-UK_PROVIDED_SYSCALLS-$(CONFIG_LIBVFSCORE) += write-3 writev-3 pwrite64-4
-UK_PROVIDED_SYSCALLS-$(CONFIG_LIBVFSCORE) += read-3 readv-3 pread64-4
-UK_PROVIDED_SYSCALLS-$(CONFIG_LIBVFSCORE) += fstat-2
-UK_PROVIDED_SYSCALLS-$(CONFIG_LIBVFSCORE) += fcntl-3
 UK_PROVIDED_SYSCALLS-$(CONFIG_LIBVFSCORE) += readlink-3
-UK_PROVIDED_SYSCALLS-$(CONFIG_LIBVFSCORE) += close-1
-UK_PROVIDED_SYSCALLS-$(CONFIG_LIBVFSCORE) += lseek-3
 UK_PROVIDED_SYSCALLS-$(CONFIG_LIBVFSCORE) += link-2
-UK_PROVIDED_SYSCALLS-$(CONFIG_LIBVFSCORE) += ftruncate-2
 UK_PROVIDED_SYSCALLS-$(CONFIG_LIBVFSCORE) += truncate-2
 UK_PROVIDED_SYSCALLS-$(CONFIG_LIBVFSCORE) += access-2
 UK_PROVIDED_SYSCALLS-$(CONFIG_LIBVFSCORE) += faccessat-4
-UK_PROVIDED_SYSCALLS-$(CONFIG_LIBVFSCORE) += fallocate-4
 UK_PROVIDED_SYSCALLS-$(CONFIG_LIBVFSCORE) += chdir-1
 UK_PROVIDED_SYSCALLS-$(CONFIG_LIBVFSCORE) += fchdir-1
 UK_PROVIDED_SYSCALLS-$(CONFIG_LIBVFSCORE) += chmod-2
 UK_PROVIDED_SYSCALLS-$(CONFIG_LIBVFSCORE) += fchmod-2
-UK_PROVIDED_SYSCALLS-$(CONFIG_LIBVFSCORE) += dup-1
-UK_PROVIDED_SYSCALLS-$(CONFIG_LIBVFSCORE) += dup3-3
-UK_PROVIDED_SYSCALLS-$(CONFIG_LIBVFSCORE) += dup2-2
-UK_PROVIDED_SYSCALLS-$(CONFIG_LIBVFSCORE) += pwritev-4
 UK_PROVIDED_SYSCALLS-$(CONFIG_LIBVFSCORE) += utime-2
 UK_PROVIDED_SYSCALLS-$(CONFIG_LIBVFSCORE) += utimes-2
 UK_PROVIDED_SYSCALLS-$(CONFIG_LIBVFSCORE) += mknod-3
 UK_PROVIDED_SYSCALLS-$(CONFIG_LIBVFSCORE) += rmdir-1
 UK_PROVIDED_SYSCALLS-$(CONFIG_LIBVFSCORE) += rename-2
-UK_PROVIDED_SYSCALLS-$(CONFIG_LIBVFSCORE) += fsync-1
-UK_PROVIDED_SYSCALLS-$(CONFIG_LIBVFSCORE) += fdatasync-1
-UK_PROVIDED_SYSCALLS-$(CONFIG_LIBVFSCORE) += preadv-4
 UK_PROVIDED_SYSCALLS-$(CONFIG_LIBVFSCORE) += umask-1
 UK_PROVIDED_SYSCALLS-$(CONFIG_LIBVFSCORE) += lstat-2
-UK_PROVIDED_SYSCALLS-$(CONFIG_LIBVFSCORE) += flock-2
 UK_PROVIDED_SYSCALLS-$(CONFIG_LIBVFSCORE) += getcwd-2
 UK_PROVIDED_SYSCALLS-$(CONFIG_LIBVFSCORE) += utimensat-4
 UK_PROVIDED_SYSCALLS-$(CONFIG_LIBVFSCORE) += futimesat-3
 UK_PROVIDED_SYSCALLS-$(CONFIG_LIBVFSCORE) += sync-0
 UK_PROVIDED_SYSCALLS-$(CONFIG_LIBVFSCORE) += mount-5
-UK_PROVIDED_SYSCALLS-$(CONFIG_LIBVFSCORE) += ioctl-3
 UK_PROVIDED_SYSCALLS-$(CONFIG_LIBVFSCORE) += statfs-2
 UK_PROVIDED_SYSCALLS-$(CONFIG_LIBVFSCORE) += fstatfs-2
 UK_PROVIDED_SYSCALLS-$(CONFIG_LIBVFSCORE) += fchown-3
