+++ lib/vfscore/exportsyms.uk
@@ -1,15 +1,15 @@
+vfscore_fdops
+vfscore_get_file
+vfscore_put_file
 vfscore_alloc_fd
 vfscore_put_fd
 vfscore_install_fd
-vfscore_get_file
-vfscore_put_file
 mount
 uk_syscall_e_mount
 uk_syscall_r_mount
 vfscore_nullop
 vfscore_release_mp_dentries
 vfscore_vget
-vfscore_uiomove
 vfscore_vop_nullop
 vfscore_vop_einval
 vfscore_vop_eperm
@@ -22,38 +22,15 @@ creat
 creat64
 uk_syscall_e_creat
 uk_syscall_r_creat
-write
-uk_syscall_e_write
-uk_syscall_r_write
-close
-uk_syscall_e_close
-uk_syscall_r_close
-read
-uk_syscall_e_read
-uk_syscall_r_read
 mkdir
 uk_syscall_e_mkdir
 uk_syscall_r_mkdir
 mkdirat
 uk_syscall_e_mkdirat
 uk_syscall_r_mkdirat
-fsync
-uk_syscall_e_fsync
-uk_syscall_r_fsync
-fstat
-fstat64
-uk_syscall_e_fstat
-uk_syscall_r_fstat
-flock
-uk_syscall_e_flock
-uk_syscall_r_flock
 fhold
 fdrop
 fget
-ftruncate
-ftruncate64
-uk_syscall_e_ftruncate
-uk_syscall_r_ftruncate
 stat
 stat64
 uk_syscall_e_stat
@@ -67,15 +44,6 @@ uk_syscall_r_fchmod
 fchown
 uk_syscall_e_fchown
 uk_syscall_r_fchown
-dup
-uk_syscall_e_dup
-uk_syscall_r_dup
-dup2
-uk_syscall_e_dup2
-uk_syscall_r_dup2
-dup3
-uk_syscall_e_dup3
-uk_syscall_r_dup3
 sync
 uk_syscall_e_sync
 uk_syscall_r_sync
@@ -127,22 +95,6 @@ readdir
 readdir_r
 readdir64
 closedir
-pread
-pread64
-uk_syscall_e_pread64
-uk_syscall_r_pread64
-pwritev
-pwritev64
-uk_syscall_e_pwritev
-uk_syscall_r_pwritev
-pwrite
-pwrite64
-uk_syscall_e_pwrite64
-uk_syscall_r_pwrite64
-readv
-uk_syscall_e_readv
-uk_syscall_r_readv
-writev
 truncate
 truncate64
 uk_syscall_e_truncate
@@ -150,16 +102,6 @@ uk_syscall_r_truncate
 mknod
 uk_syscall_e_mknod
 uk_syscall_r_mknod
-preadv
-preadv64
-uk_syscall_e_preadv
-uk_syscall_r_preadv
-ioctl
-uk_syscall_e_ioctl
-uk_syscall_r_ioctl
-fdatasync
-uk_syscall_e_fdatasync
-uk_syscall_r_fdatasync
 fdopendir
 dirfd
 rewinddir
@@ -187,16 +129,6 @@ uk_syscall_r_faccessat
 readlink
 uk_syscall_e_readlink
 uk_syscall_r_readlink
-fallocate
-fallocate64
-uk_syscall_e_fallocate
-uk_syscall_r_fallocate
-lseek
-lseek64
-uk_syscall_e_lseek
-uk_syscall_r_lseek
-uk_syscall_e_writev
-uk_syscall_r_writev
 umask
 uk_syscall_e_umask
 uk_syscall_r_umask
@@ -211,9 +143,6 @@ vput
 vref
 vflush
 dref
-fcntl
-uk_syscall_e_fcntl
-uk_syscall_r_fcntl
 readdir_r
 readdir64_r
 fstatfs
@@ -268,18 +197,9 @@ scandir
 getdents
 uk_syscall_e_getdents
 uk_syscall_r_getdents
-eventpoll_fini
-eventpoll_notify_close
-eventpoll_add
-eventpoll_add_unsafe
-eventpoll_mod
-eventpoll_del
-eventpoll_del_unsafe
-eventpoll_wait
 getdents64
 uk_syscall_e_getdents64
 uk_syscall_r_getdents64
-eventpoll_signal
 __fxstat
 __fxstat64
 __fxstatat
