+++ lib/vfscore/main.c
@@ -87,19 +87,6 @@ static inline int libc_error(int err)
     return -1;
 }
 
-// In BSD's internal implementation of read() and write() code, for example
-// sosend_generic(), a partial read or write returns both an EWOULDBLOCK error
-// *and* a non-zero number of written bytes. In that case, we need to zero the
-// error, so the system call appear a successful partial read/write.
-// In FreeBSD, dofilewrite() and dofileread() (sys_generic.c) do this too.
-static inline int has_error(int error, int bytes)
-{
-	/* TODO: OSv checks also for ERESTART */
-	return error && (
-		(bytes == 0) ||
-		(error != EWOULDBLOCK && error != EINTR));
-}
-
 static inline mode_t apply_umask(mode_t mode)
 {
 	return mode & ~ukarch_load_n(&global_umask);
@@ -143,7 +130,7 @@ UK_LLSYSCALL_R_DEFINE(int, open, const char*, pathname, int, flags,
 	if (error)
 		goto out_error;
 
-	error = fdalloc(fp, &fd);
+	error = fdtab_fdalloc(&fp->f_file, &fd);
 	if (error)
 		goto out_fput;
 	fdrop(fp);
@@ -246,43 +233,6 @@ UK_SYSCALL_DEFINE(int, creat, const char*, pathname, mode_t, mode)
 
 LFS64(creat);
 
-UK_TRACEPOINT(trace_vfs_close, "%d", int);
-UK_TRACEPOINT(trace_vfs_close_ret, "");
-UK_TRACEPOINT(trace_vfs_close_err, "%d", int);
-
-int fdclose(int fd)
-{
-	struct vfscore_file *fp;
-	int error;
-
-	fp = vfscore_get_file(fd);
-	if (!fp)
-		return EBADF;
-
-	error = vfscore_put_fd(fd);
-	if (!error)
-		fdrop(fp);
-
-	return error;
-}
-
-UK_SYSCALL_R_DEFINE(int, close, int, fd)
-{
-	int error;
-
-	trace_vfs_close(fd);
-	error = fdclose(fd);
-	if (error)
-		goto out_error;
-
-	trace_vfs_close_ret();
-	return 0;
-
-	out_error:
-	trace_vfs_close_err(error);
-	return -error;
-}
-
 UK_TRACEPOINT(trace_vfs_mknod, "\"%s\" 0%0o 0x%x", const char*, mode_t, dev_t);
 UK_TRACEPOINT(trace_vfs_mknod_ret, "");
 UK_TRACEPOINT(trace_vfs_mknod_err, "%d", int);
@@ -324,560 +274,6 @@ UK_SYSCALL_R_DEFINE(int, mknod, const char*, pathname, mode_t, mode, dev_t, dev)
 	return __xmknod_helper(0, pathname, mode, &dev);
 }
 
-UK_TRACEPOINT(trace_vfs_lseek, "%d 0x%x %d", int, off_t, int);
-UK_TRACEPOINT(trace_vfs_lseek_ret, "0x%x", off_t);
-UK_TRACEPOINT(trace_vfs_lseek_err, "%d", int);
-
-UK_SYSCALL_R_DEFINE(off_t, lseek, int, fd, off_t, offset, int, whence)
-{
-	struct vfscore_file *fp;
-	off_t org;
-	int error;
-
-	trace_vfs_lseek(fd, offset, whence);
-	error = fget(fd, &fp);
-	if (error)
-		goto out_error;
-
-	error = sys_lseek(fp, offset, whence, &org);
-	fdrop(fp);
-
-	if (error)
-		goto out_error;
-	trace_vfs_lseek_ret(org);
-	return org;
-
-	out_error:
-	trace_vfs_lseek_err(error);
-	return -error;
-}
-
-#ifdef lseek64
-#undef lseek64
-#endif
-
-LFS64(lseek);
-
-/**
- * Return:
- * = 0, Success and the nr of bytes read is returned in bytes parameter.
- * < 0, error code.
- */
-static ssize_t do_preadv(struct vfscore_file *fp, const struct iovec *iov,
-			 int iovcnt, off_t offset, ssize_t *bytes)
-{
-	size_t cnt;
-	int error;
-
-	UK_ASSERT(fp && iov);
-
-	/* Otherwise, try to read the file. */
-	error = sys_read(fp, iov, iovcnt, offset, &cnt);
-
-	if (has_error(error, cnt))
-		goto out_error;
-
-	*bytes = cnt;
-	return 0;
-
-out_error:
-	return -error;
-}
-
-UK_TRACEPOINT(trace_vfs_preadv, "%d %p 0x%x 0x%x", int, const struct iovec*,
-	      int, off_t);
-UK_TRACEPOINT(trace_vfs_preadv_ret, "0x%x", ssize_t);
-UK_TRACEPOINT(trace_vfs_preadv_err, "%d", int);
-
-UK_SYSCALL_R_DEFINE(ssize_t, preadv, int, fd, const struct iovec*, iov,
-	int, iovcnt, off_t, offset)
-{
-	struct vfscore_file *fp;
-	ssize_t bytes;
-	int error;
-
-	trace_vfs_preadv(fd, iov, iovcnt, offset);
-	error = fget(fd, &fp);
-	if (error) {
-		error = -error;
-		goto out_error;
-	}
-
-	/* Check if the file is indeed seekable. */
-	if (fp->f_vfs_flags & UK_VFSCORE_NOPOS) {
-		error = -ESPIPE;
-		goto out_error_fdrop;
-	}
-	/* Check if the file has not already been read and that
-	 * is not a character device.
-	 */
-	else if (fp->f_offset < 0 &&
-		(fp->f_dentry == NULL ||
-		 fp->f_dentry->d_vnode->v_type != VCHR)) {
-		error = -EINVAL;
-		goto out_error_fdrop;
-	}
-
-	/* Otherwise, try to read the file. */
-	error = do_preadv(fp, iov, iovcnt, offset, &bytes);
-
-out_error_fdrop:
-	fdrop(fp);
-
-	if (error < 0)
-		goto out_error;
-
-	trace_vfs_preadv_ret(bytes);
-	return bytes;
-
-out_error:
-	trace_vfs_preadv_err(error);
-	return error;
-}
-
-#ifdef preadv64
-#undef preadv64
-#endif
-
-LFS64(preadv);
-
-UK_TRACEPOINT(trace_vfs_pread, "%d %p 0x%x 0x%x", int, void*, size_t, off_t);
-UK_TRACEPOINT(trace_vfs_pread_ret, "0x%x", ssize_t);
-UK_TRACEPOINT(trace_vfs_pread_err, "%d", int);
-
-/*
- * Some libc's define some macros that remove the 64 suffix
- * from some system call function names (e.g., <unistd.h>, <fcntl.h>).
- * We need to undefine them here so that our system call
- * registration does not fail in such a case.
- */
-#ifdef pread64
-#undef pread64
-#endif
-
-UK_LLSYSCALL_R_DEFINE(ssize_t, pread64, int, fd,
-		      void *, buf, size_t, count, off_t, offset)
-{
-	trace_vfs_pread(fd, buf, count, offset);
-	struct iovec iov = {
-			.iov_base	= buf,
-			.iov_len	= count,
-	};
-	ssize_t bytes;
-
-	bytes = uk_syscall_r_preadv((long) fd, (long) &iov,
-				    1, (long) offset);
-	if (bytes < 0)
-		trace_vfs_pread_err(bytes);
-	else
-		trace_vfs_pread_ret(bytes);
-	return bytes;
-}
-
-#if UK_LIBC_SYSCALLS
-ssize_t pread(int fd, void *buf, size_t count, off_t offset)
-{
-	return uk_syscall_e_pread64((long) fd, (long) buf,
-				    (long) count, (long) offset);
-}
-
-LFS64(pread);
-#endif /* UK_LIBC_SYSCALLS */
-
-UK_TRACEPOINT(trace_vfs_readv, "%d %p 0x%x", int, const struct iovec*,
-	      int);
-UK_TRACEPOINT(trace_vfs_readv_ret, "0x%x", ssize_t);
-UK_TRACEPOINT(trace_vfs_readv_err, "%d", int);
-
-UK_SYSCALL_R_DEFINE(ssize_t, readv,
-		  int, fd, const struct iovec *, iov, int, iovcnt)
-{
-	struct vfscore_file *fp;
-	ssize_t bytes;
-	int error;
-
-	trace_vfs_readv(fd, iov, iovcnt);
-	error = fget(fd, &fp);
-	if (error) {
-		error = -error;
-		goto out_error;
-	}
-
-	/* Check if the file has not already been read and that is
-	 * not a character device.
-	 */
-	if (fp->f_offset < 0 &&
-	   (fp->f_dentry == NULL ||
-	    fp->f_dentry->d_vnode->v_type != VCHR)) {
-		error = -EINVAL;
-		goto out_error_fdrop;
-	}
-
-	/* Otherwise, try to read the file. */
-	error = do_preadv(fp, iov, iovcnt, -1, &bytes);
-
-out_error_fdrop:
-	fdrop(fp);
-
-	if (error < 0)
-		goto out_error;
-
-	trace_vfs_readv_ret(bytes);
-	return bytes;
-
-out_error:
-	trace_vfs_readv_err(error);
-	return error;
-}
-
-UK_TRACEPOINT(trace_vfs_read, "%d %p %d", int, void *, int);
-UK_TRACEPOINT(trace_vfs_read_ret, "0x%x", ssize_t);
-UK_TRACEPOINT(trace_vfs_read_err, "%d", int);
-
-UK_SYSCALL_R_DEFINE(ssize_t, read, int, fd, void *, buf, size_t, count)
-{
-	ssize_t bytes;
-
-	UK_ASSERT(buf);
-
-	struct iovec iov = {
-			.iov_base	= buf,
-			.iov_len	= count,
-	};
-
-	trace_vfs_read(fd, buf, count);
-
-	bytes = uk_syscall_r_readv((long) fd, (long) &iov, 1);
-	if (bytes < 0)
-		trace_vfs_read_err(bytes);
-	else
-		trace_vfs_read_ret(bytes);
-	return bytes;
-}
-
-static int do_pwritev(struct vfscore_file *fp, const struct iovec *iov,
-		      int iovcnt, off_t offset, ssize_t *bytes)
-{
-	int error;
-	size_t cnt;
-
-	UK_ASSERT(bytes);
-
-	/* Otherwise, try to read the file. */
-	error = sys_write(fp, iov, iovcnt, offset, &cnt);
-
-	if (has_error(error, cnt))
-		goto out_error;
-
-	*bytes = cnt;
-	return 0;
-
-out_error:
-	*bytes = -1;
-	return -error;
-}
-
-UK_TRACEPOINT(trace_vfs_pwritev, "%d %p 0x%x 0x%x", int, const struct iovec*,
-	      int, off_t);
-UK_TRACEPOINT(trace_vfs_pwritev_ret, "0x%x", ssize_t);
-UK_TRACEPOINT(trace_vfs_pwritev_err, "%d", int);
-
-UK_SYSCALL_R_DEFINE(ssize_t, pwritev, int, fd, const struct iovec*, iov,
-			int, iovcnt, off_t, offset)
-{
-	struct vfscore_file *fp;
-	ssize_t bytes;
-	int error;
-
-	trace_vfs_pwritev(fd, iov, iovcnt, offset);
-	error = fget(fd, &fp);
-	if (error) {
-		error = -error;
-		goto out_error;
-	}
-
-	/* Check if the file is indeed seekable. */
-	if (fp->f_vfs_flags & UK_VFSCORE_NOPOS) {
-		error = -ESPIPE;
-		goto out_error_fdrop;
-	}
-	/* Check if the file has not already been written to and that it is
-	 * not a character device.
-	 */
-	else if (fp->f_offset < 0 &&
-		(fp->f_dentry == NULL ||
-		 fp->f_dentry->d_vnode->v_type != VCHR)) {
-		error = -EINVAL;
-		goto out_error_fdrop;
-	}
-
-	/* Otherwise, try to read the file. */
-	error = do_pwritev(fp, iov, iovcnt, offset, &bytes);
-
-out_error_fdrop:
-	fdrop(fp);
-
-	if (error < 0)
-		goto out_error;
-
-	trace_vfs_pwritev_ret(bytes);
-	return bytes;
-
-out_error:
-	trace_vfs_pwritev_err(error);
-	return error;
-}
-
-#ifdef pwritev64
-#undef pwritev64
-#endif
-
-LFS64(pwritev);
-
-UK_TRACEPOINT(trace_vfs_pwrite, "%d %p 0x%x 0x%x", int, const void*, size_t,
-	      off_t);
-UK_TRACEPOINT(trace_vfs_pwrite_ret, "0x%x", ssize_t);
-UK_TRACEPOINT(trace_vfs_pwrite_err, "%d", int);
-
-/*
- * Some libc's define some macros that remove the 64 suffix
- * from some system call function names (e.g., <unistd.h>, <fcntl.h>).
- * We need to undefine them here so that our system call
- * registration does not fail in such a case.
- */
-#ifdef pwrite64
-#undef pwrite64
-#endif
-
-UK_LLSYSCALL_R_DEFINE(ssize_t, pwrite64, int, fd,
-		      const void *, buf, size_t, count, off_t, offset)
-{
-	trace_vfs_pwrite(fd, buf, count, offset);
-	struct iovec iov = {
-			.iov_base	= (void *)buf,
-			.iov_len	= count,
-	};
-	ssize_t bytes;
-
-	bytes = uk_syscall_r_pwritev((long) fd, (long) &iov,
-				     1, (long) offset);
-	if (bytes < 0)
-		trace_vfs_pwrite_err(bytes);
-	else
-		trace_vfs_pwrite_ret(bytes);
-	return bytes;
-}
-
-#if UK_LIBC_SYSCALLS
-ssize_t pwrite(int fd, const void *buf, size_t count, off_t offset)
-{
-	return uk_syscall_e_pwrite64((long) fd, (long) buf,
-				     (long) count, (long) offset);
-}
-
-LFS64(pwrite);
-#endif /* UK_LIBC_SYSCALLS */
-
-UK_TRACEPOINT(trace_vfs_writev, "%d %p 0x%x 0x%x", int, const struct iovec*,
-	      int);
-UK_TRACEPOINT(trace_vfs_writev_ret, "0x%x", ssize_t);
-UK_TRACEPOINT(trace_vfs_writev_err, "%d", int);
-
-UK_SYSCALL_R_DEFINE(ssize_t, writev,
-		  int, fd, const struct iovec *, vec, int, vlen)
-{
-	struct vfscore_file *fp;
-	ssize_t bytes;
-	int error;
-
-	trace_vfs_writev(fd, vec, vlen);
-	error = fget(fd, &fp);
-	if (error) {
-		error = -error;
-		goto out_error;
-	}
-
-	/* Check if the file has not already been written to and
-	 * that it is not a character device.
-	 */
-	if (fp->f_offset < 0 &&
-	   (fp->f_dentry == NULL ||
-	    fp->f_dentry->d_vnode->v_type != VCHR)) {
-		error = -EINVAL;
-		goto out_error_fdrop;
-	}
-
-	/* Otherwise, try to read the file. */
-	error = do_pwritev(fp, vec, vlen, -1, &bytes);
-
-out_error_fdrop:
-	fdrop(fp);
-
-	if (error < 0)
-		goto out_error;
-
-	trace_vfs_pwritev_ret(bytes);
-	return bytes;
-
-out_error:
-	trace_vfs_pwritev_err(error);
-	return error;
-}
-
-UK_TRACEPOINT(trace_vfs_write, "%d %p 0x%x 0x%x", int, const void *,
-	      size_t);
-UK_TRACEPOINT(trace_vfs_write_ret, "0x%x", ssize_t);
-UK_TRACEPOINT(trace_vfs_write_err, "%d", int);
-
-UK_SYSCALL_R_DEFINE(ssize_t, write, int, fd, const void *, buf, size_t, count)
-{
-	ssize_t bytes;
-
-	UK_ASSERT(buf);
-
-	struct iovec iov = {
-			.iov_base	= (void *)buf,
-			.iov_len	= count,
-	};
-	trace_vfs_write(fd, buf, count);
-	bytes = uk_syscall_r_writev((long) fd, (long) &iov, 1);
-	if (bytes < 0)
-		trace_vfs_write_err(errno);
-	else
-		trace_vfs_write_ret(bytes);
-	return bytes;
-}
-
-UK_TRACEPOINT(trace_vfs_ioctl, "%d 0x%x", int, unsigned long);
-UK_TRACEPOINT(trace_vfs_ioctl_ret, "");
-UK_TRACEPOINT(trace_vfs_ioctl_err, "%d", int);
-
-UK_LLSYSCALL_R_DEFINE(int, ioctl, int, fd, unsigned long int, request,
-		void*, arg)
-{
-	struct vfscore_file *fp;
-	int error;
-
-	trace_vfs_ioctl(fd, request);
-	/* glibc ABI provides a variadic prototype for ioctl so we need to agree
-	 * with it, since we now include sys/ioctl.h
-	 * read the first argument and pass it to sys_ioctl() */
-
-	error = fget(fd, &fp);
-	if (error)
-		goto out_errno;
-
-	error = sys_ioctl(fp, request, arg);
-	fdrop(fp);
-
-	if (error)
-		goto out_errno;
-	trace_vfs_ioctl_ret();
-	return 0;
-
-out_errno:
-	trace_vfs_ioctl_err(error);
-	return -error;
-}
-
-#if UK_LIBC_SYSCALLS
-int ioctl(int fd, unsigned long int request, ...)
-{
-	va_list ap;
-	void *arg;
-
-	va_start(ap, request);
-	arg = va_arg(ap, void*);
-	va_end(ap);
-
-	return uk_syscall_e_ioctl((long) fd, (long) request, (long) arg);
-}
-#endif /* UK_LIBC_SYSCALLS */
-
-UK_TRACEPOINT(trace_vfs_fsync, "%d", int);
-UK_TRACEPOINT(trace_vfs_fsync_ret, "");
-UK_TRACEPOINT(trace_vfs_fsync_err, "%d", int);
-
-UK_SYSCALL_R_DEFINE(int, fsync, int, fd)
-{
-	struct vfscore_file *fp;
-	int error;
-
-	trace_vfs_fsync(fd);
-	error = fget(fd, &fp);
-	if (error)
-		goto out_error;
-
-	error = sys_fsync(fp);
-	fdrop(fp);
-
-	if (error)
-		goto out_error;
-	trace_vfs_fsync_ret();
-	return 0;
-
-	out_error:
-	trace_vfs_fsync_err(error);
-	return -error;
-}
-
-UK_SYSCALL_R_DEFINE(int, fdatasync, int, fd)
-{
-	// TODO: See if we can do less than fsync().
-	return fsync(fd);
-}
-
-UK_TRACEPOINT(trace_vfs_fstat, "%d %p", int, struct stat*);
-UK_TRACEPOINT(trace_vfs_fstat_ret, "");
-UK_TRACEPOINT(trace_vfs_fstat_err, "%d", int);
-
-static int __fxstat_helper(int ver __unused, int fd, struct stat *st)
-{
-	struct vfscore_file *fp;
-	int error;
-
-	trace_vfs_fstat(fd, st);
-
-	error = fget(fd, &fp);
-	if (error)
-		goto out_errno;
-
-	error = sys_fstat(fp, st);
-	fdrop(fp);
-
-	if (error)
-		goto out_errno;
-	trace_vfs_fstat_ret();
-	return 0;
-
-	out_errno:
-	trace_vfs_fstat_err(error);
-	errno = error;
-	return -1;
-}
-
-#if UK_LIBC_SYSCALLS
-int __fxstat(int ver __unused, int fd, struct stat *st)
-{
-	return __fxstat_helper(ver, fd, st);
-}
-#ifdef __fxstat64
-#undef __fxstat64
-#endif
-
-LFS64(__fxstat);
-#endif /* UK_LIBC_SYSCALLS */
-
-UK_SYSCALL_DEFINE(int, fstat, int, fd, struct stat *, st)
-{
-	return __fxstat_helper(1, fd, st);
-}
-
-#ifdef fstat64
-#undef fstat64
-#endif
-
-LFS64(fstat);
 
 static int __fxstatat_helper(int ver __unused, int dirfd, const char *pathname,
 		struct stat *st, int flags);
@@ -914,33 +310,6 @@ UK_SYSCALL_R_DEFINE(int, newfstatat, int, dirfd, const char*, path,
 				struct stat*, st, int, flags)
 {
 	return __fxstatat_helper(1, dirfd, path, st, flags);
-}
-
-UK_SYSCALL_R_DEFINE(int, flock, int, fd, int, operation)
-{
-	struct vfscore_file *file;
-	int error;
-
-	error = fget(fd, &file);
-	if (error)
-		goto out_error;
-
-	switch (operation) {
-	case LOCK_SH:
-	case LOCK_SH | LOCK_NB:
-	case LOCK_EX:
-	case LOCK_EX | LOCK_NB:
-	case LOCK_UN:
-		break;
-	default:
-		error = EINVAL;
-		goto out_error;
-	}
-
-	return 0;
-
-	out_error:
-	return -error;
 }
 
 UK_TRACEPOINT(trace_vfs_readdir, "%d %p", int, struct dirent*);
@@ -1893,223 +1262,6 @@ out_error:
 	return ERR2PTR(-error);
 }
 
-UK_TRACEPOINT(trace_vfs_dup, "%d", int);
-UK_TRACEPOINT(trace_vfs_dup_ret, "\"%s\"", int);
-UK_TRACEPOINT(trace_vfs_dup_err, "%d", int);
-/*
- * Duplicate a file descriptor
- */
-UK_SYSCALL_R_DEFINE(int, dup, int, oldfd)
-{
-	struct vfscore_file *fp;
-	int newfd;
-	int error;
-
-	trace_vfs_dup(oldfd);
-	error = fget(oldfd, &fp);
-	if (error)
-		goto out_error;
-
-	error = fdalloc(fp, &newfd);
-	if (error)
-		goto out_fdrop;
-
-	fdrop(fp);
-	trace_vfs_dup_ret(newfd);
-	return newfd;
-
-	out_fdrop:
-	fdrop(fp);
-
-	out_error:
-	trace_vfs_dup_err(error);
-	if (error > 0)
-		return -error;
-	return error;
-}
-
-UK_TRACEPOINT(trace_vfs_dup3, "%d %d 0x%x", int, int, int);
-UK_TRACEPOINT(trace_vfs_dup3_ret, "%d", int);
-UK_TRACEPOINT(trace_vfs_dup3_err, "%d", int);
-/*
- * Duplicate a file descriptor to a particular value.
- */
-UK_SYSCALL_R_DEFINE(int, dup3, int, oldfd, int, newfd, int, flags)
-{
-	struct vfscore_file *fp, *fp_new;
-	int error;
-
-	trace_vfs_dup3(oldfd, newfd, flags);
-	/*
-	 * Don't allow any argument but O_CLOEXEC.  But we even ignore
-	 * that as we don't support exec() and thus don't care.
-	 */
-	if ((flags & ~O_CLOEXEC) != 0) {
-		error = EINVAL;
-		goto out_error;
-	}
-
-	if (oldfd == newfd) {
-		error = EINVAL;
-		goto out_error;
-	}
-
-	error = fget(oldfd, &fp);
-	if (error)
-		goto out_error;
-
-	error = fget(newfd, &fp_new);
-	if (error == 0) {
-		/* if newfd is open, then close it */
-		error = close(newfd);
-		if (error)
-			goto out_error;
-	}
-
-	error = vfscore_reserve_fd(newfd);
-	if (error)
-		goto out_error;
-
-	error = vfscore_install_fd(newfd, fp);
-	if (error) {
-		fdrop(fp);
-		goto out_error;
-	}
-
-	fdrop(fp);
-	trace_vfs_dup3_ret(newfd);
-	return newfd;
-
-	out_error:
-	trace_vfs_dup3_err(error);
-	if(error > 0)
-		return -error;
-	return error;
-}
-
-UK_SYSCALL_R_DEFINE(int, dup2, int, oldfd, int, newfd)
-{
-	if (oldfd == newfd)
-		return newfd;
-
-	return uk_syscall_r_dup3(oldfd, newfd, 0);
-}
-
-/*
- * The file control system call.
- */
-#define SETFL (O_APPEND | O_ASYNC | O_DIRECT | O_NOATIME | O_NONBLOCK)
-
-UK_TRACEPOINT(trace_vfs_fcntl, "%d %d 0x%x", int, int, int);
-UK_TRACEPOINT(trace_vfs_fcntl_ret, "\"%s\"", int);
-UK_TRACEPOINT(trace_vfs_fcntl_err, "%d", int);
-
-UK_LLSYSCALL_R_DEFINE(int, fcntl, int, fd, unsigned int, cmd, int, arg)
-{
-	struct vfscore_file *fp;
-	int ret = 0, error;
-#if defined(FIONBIO) && defined(FIOASYNC)
-	int tmp;
-#endif
-
-	trace_vfs_fcntl(fd, cmd, arg);
-	error = fget(fd, &fp);
-	if (error)
-		goto out_errno;
-
-	// An important note about our handling of FD_CLOEXEC / O_CLOEXEC:
-	// close-on-exec shouldn't have been a file flag (fp->f_flags) - it is a
-	// file descriptor flag, meaning that that two dup()ed file descriptors
-	// could have different values for FD_CLOEXEC. Our current implementation
-	// *wrongly* makes close-on-exec an f_flag (using the bit O_CLOEXEC).
-	// There is little practical difference, though, because this flag is
-	// ignored in OSv anyway, as it doesn't support exec().
-	switch (cmd) {
-	case F_DUPFD:
-		error = fdalloc(fp, &ret);
-		if (error)
-			goto out_errno;
-		break;
-	case F_GETFD:
-		ret = (fp->f_flags & O_CLOEXEC) ? FD_CLOEXEC : 0;
-		break;
-	case F_SETFD:
-		FD_LOCK(fp);
-		fp->f_flags = (fp->f_flags & ~O_CLOEXEC) |
-				((arg & FD_CLOEXEC) ? O_CLOEXEC : 0);
-		FD_UNLOCK(fp);
-		break;
-	case F_GETFL:
-		// As explained above, the O_CLOEXEC should have been in f_flags,
-		// and shouldn't be returned. Linux always returns 0100000 ("the
-		// flag formerly known as O_LARGEFILE) so let's do it too.
-		ret = (vfscore_oflags(fp->f_flags) & ~O_CLOEXEC) | 0100000;
-		break;
-	case F_SETFL:
-		FD_LOCK(fp);
-		fp->f_flags = vfscore_fflags((vfscore_oflags(fp->f_flags) & ~SETFL) |
-				(arg & SETFL));
-		FD_UNLOCK(fp);
-
-#if defined(FIONBIO) && defined(FIOASYNC)
-		/* Sync nonblocking/async state with file flags */
-		tmp = fp->f_flags & FNONBLOCK;
-		vfs_ioctl(fp, FIONBIO, &tmp);
-		tmp = fp->f_flags & FASYNC;
-		vfs_ioctl(fp, FIOASYNC, &tmp);
-#endif
-		break;
-	case F_DUPFD_CLOEXEC:
-		error = fdalloc(fp, &ret);
-		if (error)
-			goto out_errno;
-		fp->f_flags |= O_CLOEXEC;
-		break;
-	case F_SETLK:
-		uk_pr_warn_once("fcntl(F_SETLK) stubbed\n");
-		break;
-	case F_GETLK:
-		uk_pr_warn_once("fcntl(F_GETLK) stubbed\n");
-		break;
-	case F_SETLKW:
-		uk_pr_warn_once("fcntl(F_SETLKW) stubbed\n");
-		break;
-	case F_SETOWN:
-		uk_pr_warn_once("fcntl(F_SETOWN) stubbed\n");
-		break;
-	default:
-		uk_pr_err("unsupported fcntl cmd 0x%x\n", cmd);
-		error = EINVAL;
-	}
-
-	fdrop(fp);
-	if (error)
-		goto out_errno;
-	trace_vfs_fcntl_ret(ret);
-	return ret;
-
-out_errno:
-	trace_vfs_fcntl_err(error);
-	return -error;
-}
-
-#if UK_LIBC_SYSCALLS
-int fcntl(int fd, int cmd, ...)
-{
-	int arg = 0;
-	va_list ap;
-
-	va_start(ap, cmd);
-	if (cmd == F_SETFD ||
-	    cmd == F_SETFL) {
-		arg = va_arg(ap, int);
-	}
-	va_end(ap);
-
-	return uk_syscall_e_fcntl(fd, cmd, arg);
-}
-#endif /* UK_LIBC_SYSCALLS */
-
 UK_TRACEPOINT(trace_vfs_access, "\"%s\" 0%0o", const char*, int);
 UK_TRACEPOINT(trace_vfs_access_ret, "");
 UK_TRACEPOINT(trace_vfs_access_err, "%d", int);
@@ -2251,39 +1403,6 @@ UK_SYSCALL_R_DEFINE(int, truncate, const char*, pathname, off_t, length)
 
 LFS64(truncate);
 
-UK_TRACEPOINT(trace_vfs_ftruncate, "%d 0x%x", int, off_t);
-UK_TRACEPOINT(trace_vfs_ftruncate_ret, "");
-UK_TRACEPOINT(trace_vfs_ftruncate_err, "%d", int);
-
-UK_SYSCALL_R_DEFINE(int, ftruncate, int, fd, off_t, length)
-{
-	trace_vfs_ftruncate(fd, length);
-	struct vfscore_file *fp;
-	int error;
-
-	error = fget(fd, &fp);
-	if (error)
-		goto out_error;
-
-	error = sys_ftruncate(fp, length);
-	fdrop(fp);
-
-	if (error)
-		goto out_error;
-	trace_vfs_ftruncate_ret();
-	return 0;
-
-	out_error:
-	trace_vfs_ftruncate_err(error);
-	return -error;
-}
-
-#ifdef ftruncate64
-#undef ftruncate64
-#endif
-
-LFS64(ftruncate);
-
 UK_SYSCALL_DEFINE(ssize_t, readlink, const char *, pathname, char *, buf, size_t, bufsize)
 {
 	struct task *t = main_task;
@@ -2313,39 +1432,6 @@ UK_SYSCALL_DEFINE(ssize_t, readlink, const char *, pathname, char *, buf, size_t
 	errno = error;
 	return -1;
 }
-
-UK_TRACEPOINT(trace_vfs_fallocate, "%d %d 0x%x 0x%x", int, int, loff_t, loff_t);
-UK_TRACEPOINT(trace_vfs_fallocate_ret, "");
-UK_TRACEPOINT(trace_vfs_fallocate_err, "%d", int);
-
-UK_SYSCALL_R_DEFINE(int, fallocate, int, fd, int, mode, loff_t, offset, loff_t, len)
-{
-	struct vfscore_file *fp;
-	int error;
-
-	trace_vfs_fallocate(fd, mode, offset, len);
-	error = fget(fd, &fp);
-	if (error)
-		goto out_error;
-
-	error = sys_fallocate(fp, mode, offset, len);
-	fdrop(fp);
-
-	if (error)
-		goto out_error;
-	trace_vfs_fallocate_ret();
-	return 0;
-
-	out_error:
-	trace_vfs_fallocate_err(error);
-	return -error;
-}
-
-#ifdef fallocate64
-#undef fallocate64
-#endif
-
-LFS64(fallocate);
 
 UK_TRACEPOINT(trace_vfs_utimes, "\"%s\"", const char*);
 UK_TRACEPOINT(trace_vfs_utimes_ret, "");
